pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fhir2dicom4ortho"
version = "0.1.2"
description = "A FHIR API to dicom4ortho"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT License"}
authors = [
    {name = "Toni Magni", email = "amagni@open-ortho.org"},
]
keywords = ["DICOM", "VL", "Visible Light Photography", "Orthodontics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Image Processing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3"
]
dependencies = [
    "requests",
    "sqlalchemy",
    "apscheduler",
    "pydantic",
    "fhir.resources",
    "httpx",
    "fastapi",
    "argparse",
    "pynetdicom",
    "pydicom<=2.5",
    "pillow",
    "prettytable",
    "numpy",
    "uvicorn",
    "dicom4ortho"
]

[project.urls]
homepage = 'https://github.com/open-ortho/fhir2dicom4ortho'

[project.scripts]
fhir2dicom4ortho = "fhir2dicom4ortho.entry_points:fhir_api"

[tool.setuptools.packages.find]
include = ["fhir2dicom4ortho*"]

[tool.poetry.group.dev.dependencies]
toml = "^0.10.2"
invoke = "^1.5.0"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.organization]
name = "open-ortho"
url = "https://open-ortho.org"